Fix Alarm issue, Refactor filename
