+++ app/src/main/java/xyz/bcfriends/dra/util/DeviceBootReceiver.java
@@ -6,6 +6,7 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import androidx.preference.PreferenceManager;
+import xyz.bcfriends.dra.DailyAlarmPresenter;
 import xyz.bcfriends.dra.DailyAlarmReceiver;
 
 import java.util.Objects;
@@ -18,7 +19,7 @@ public class DeviceBootReceiver extends BroadcastReceiver {
             SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
             Intent alarmIntent = new Intent(context, DailyAlarmReceiver.class);
             PendingIntent pendingIntent = PendingIntent.getBroadcast(context, 0, alarmIntent, PendingIntent.FLAG_UPDATE_CURRENT);
-            AlarmPresenter presenter = new AlarmPresenter(context, prefs, pendingIntent);
+            DailyAlarmPresenter presenter = new DailyAlarmPresenter(context, prefs, pendingIntent);
             presenter.scheduleAlarm();
         }
     }
